pyproject.toml
[project]
name = "nuts"
version = "0.1.0"
description = ""
authors = [
    {name = "ant100pshnik",email = "mypochtav@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4"
dependencies = [
    "wagtail (>=7.2.1,<8.0.0)",
    "wagtailmedia (>=0.17.2,<0.18.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "wagtail-modeltranslation (>=0.16.0,<0.17.0)",
    "django-vite (>=3.1.0,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
